+++ include/013_2DGI/GI2D/GI2DPhysics.h
@@ -89,7 +89,8 @@ struct GI2DPhysics
 		vec4 R_old;
 
 		i64vec2 cm_work;
-		i64vec4 Apq_work;
+		
+		vec4 Apq_work;
 	};
 
 #define RBP_FLAG_ACTIVE (1<<0)
